# transcendence-app/swagger-spec.yaml
openapi: 3.0.0
paths:
  /api/v1/auth/login:
    get:
      operationId: AuthController_oauth42Login
      parameters: []
      responses:
        "200":
          description: Login successfully
        "302":
          description: Redirect to 42 OAuth server
        "503":
          description: Service unavailable
      tags:
        &a1
        - auth
  /api/v1/auth/logout:
    delete:
      operationId: AuthController_logout
      parameters: []
      responses:
        "200":
          description: Logout successfully
        "403":
          description: Forbidden
        "404":
          description: Not Found
      tags: *a1
  /api/v1/auth/local/register:
    post:
      operationId: AuthController_registerLocalUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
        "422":
          description: Unprocessable Entity
      tags: *a1
  /api/v1/auth/local/login:
    post:
      operationId: AuthController_loginLocalUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Bad Request
        "403":
          description: Forbidden
      tags: *a1
  /api/v1/users:
    post:
      operationId: UserController_addUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDto"
      responses:
        "201":
          description: Create a user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
        "403":
          &a2
          description: Forbidden
        "422":
          description: Unprocessable entity
      tags:
        &a3
        - users
    patch:
      operationId: UserController_updateCurrentUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
        "403": *a2
        "422":
          description: Unprocessable Entity
      tags: *a3
    get:
      operationId: UserController_getUsers
      parameters:
        - name: limit
          required: false
          in: query
          description: The number of results (max 20)
          schema:
            minimum: 0
            maximum: 20
            type: number
        - name: offset
          required: false
          in: query
          schema:
            minimum: 0
            type: number
        - name: sort
          required: false
          in: query
          schema:
            enum:
              - "true"
              - "false"
            type: string
        - name: search
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lists all users (max 20)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
        "403": *a2
        "503":
          description: Service unavailable
      tags: *a3
  /api/v1/users/me:
    get:
      operationId: UserController_getCurrentUser
      parameters: []
      responses:
        "200":
          description: Get the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403": *a2
      tags: *a3
  "/api/v1/users/avatars/{avatarId}":
    get:
      operationId: UserController_getAvatarByAvatarId
      parameters:
        - name: avatarId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            image/jpeg:
              schema:
                type: file
                format: binary
        "400":
          description: Bad Request
        "403": *a2
        "404":
          description: Not Found
      tags: *a3
  "/api/v1/users/{userId}":
    get:
      operationId: UserController_getUserById
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Get a user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
        "403": *a2
        "404":
          description: Not Found
      tags: *a3
  /api/v1/users/avatar:
    put:
      operationId: UserController_uploadAvatar
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: ""
          content:
            image/jpeg:
              schema:
                type: file
                format: binary
        "403": *a2
        "413":
          description: Payload Too Large
        "422":
          description: Unprocessable Entity
        "503":
          description: Service Unavailable
      tags: *a3
  /api/v1/Chats:
    post:
      operationId: ChatController_addChat
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatDto"
      responses:
        "201":
          description: Create a chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "400":
          description: Bad Request
        "403":
          &a4
          description: Forbidden
        "422":
          description: Unprocessable entity
      tags:
        &a5
        - Chats
    patch:
      operationId: ChatController_updateCurrentChat
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChatDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "400":
          description: Bad Request
        "403": *a4
        "422":
          description: Unprocessable Entity
      tags: *a5
    get:
      operationId: ChatController_getChats
      parameters:
        - name: limit
          required: false
          in: query
          description: The number of results (max 20)
          schema:
            minimum: 0
            maximum: 20
            type: number
        - name: offset
          required: false
          in: query
          schema:
            minimum: 0
            type: number
        - name: sort
          required: false
          in: query
          schema:
            enum:
              - "true"
              - "false"
            type: string
        - name: search
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lists all Chats (max 20)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
        "400":
          description: Bad Request
        "403": *a4
        "503":
          description: Service unavailable
      tags: *a5
  "/api/v1/Chats/avatars/{avatarId}":
    get:
      operationId: ChatController_getAvatarByAvatarId
      parameters:
        - name: avatarId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            image/jpeg:
              schema:
                type: file
                format: binary
        "400":
          description: Bad Request
        "403": *a4
        "404":
          description: Not Found
      tags: *a5
  "/api/v1/Chats/{chatId}":
    get:
      operationId: ChatController_getChatById
      parameters:
        - name: chatId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Get a chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "400":
          description: Bad Request
        "403": *a4
        "404":
          description: Not Found
      tags: *a5
  /api/v1/Chats/avatar:
    put:
      operationId: ChatController_uploadAvatar
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: ""
          content:
            image/jpeg:
              schema:
                type: file
                format: binary
        "403": *a4
        "413":
          description: Payload Too Large
        "422":
          description: Unprocessable Entity
        "503":
          description: Service Unavailable
      tags: *a5
info:
  title: transcendence-app
  description: The transcendence-app API description
  version: "1.0"
  contact: {}
tags: []
servers: []
components:
  schemas:
    RegisterUserDto:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        confirmationPassword:
          type: string
      required:
        - username
        - fullName
        - email
        - password
        - confirmationPassword
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        password:
          type: string
          nullable: true
        avatarId:
          type: string
          nullable: true
        avatarX:
          type: number
          default: 0
        avatarY:
          type: number
          default: 0
        id:
          type: string
        createdAt:
          format: date-time
          type: string
      required:
        - username
        - email
        - fullName
        - password
        - avatarId
        - avatarX
        - avatarY
        - id
        - createdAt
    LoginUserDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    UserDto:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        password:
          type: string
          nullable: true
        avatarId:
          type: string
          nullable: true
      required:
        - username
        - email
        - fullName
        - password
        - avatarId
    UpdateUserDto:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        avatarX:
          type: number
        avatarY:
          type: number
    ChatDto:
      type: object
      properties:
        chatName:
          type: string
        password:
          type: string
          nullable: true
        avatarId:
          type: string
          nullable: true
        owner:
          type: string
      required:
        - chatName
        - password
        - avatarId
        - owner
    Chat:
      type: object
      properties:
        id:
          type: string
        chatName:
          type: string
        password:
          type: object
        avatarId:
          type: string
          nullable: true
        avatarX:
          type: number
          default: 0
        avatarY:
          type: number
          default: 0
        createdAt:
          format: date-time
          type: string
        owner:
          type: string
      required:
        - id
        - chatName
        - password
        - avatarId
        - avatarX
        - avatarY
        - createdAt
        - owner
    UpdateChatDto:
      type: object
      properties:
        chatName:
          type: string
        avatarX:
          type: number
        avatarY:
          type: number