# webapp/src/shared/generated/models/Chat.ts
/* tslint:disable */
/* eslint-disable */
/**
 * transcendence-app
 * The transcendence-app API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Chat
 */
export interface Chat {
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    chatName: string;
    /**
     * 
     * @type {object}
     * @memberof Chat
     */
    password: object;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    avatarId: string | null;
    /**
     * 
     * @type {number}
     * @memberof Chat
     */
    avatarX: number;
    /**
     * 
     * @type {number}
     * @memberof Chat
     */
    avatarY: number;
    /**
     * 
     * @type {Date}
     * @memberof Chat
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    owner: string;
}

/**
 * Check if a given object implements the Chat interface.
 */
export function instanceOfChat(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "chatName" in value;
    isInstance = isInstance && "password" in value;
    isInstance = isInstance && "avatarId" in value;
    isInstance = isInstance && "avatarX" in value;
    isInstance = isInstance && "avatarY" in value;
    isInstance = isInstance && "createdAt" in value;
    isInstance = isInstance && "owner" in value;

    return isInstance;
}

export function ChatFromJSON(json: any): Chat {
    return ChatFromJSONTyped(json, false);
}

export function ChatFromJSONTyped(json: any, ignoreDiscriminator: boolean): Chat {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'chatName': json['chatName'],
        'password': json['password'],
        'avatarId': json['avatarId'],
        'avatarX': json['avatarX'],
        'avatarY': json['avatarY'],
        'createdAt': (new Date(json['createdAt'])),
        'owner': json['owner'],
    };
}

export function ChatToJSON(value?: Chat | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'chatName': value.chatName,
        'password': value.password,
        'avatarId': value.avatarId,
        'avatarX': value.avatarX,
        'avatarY': value.avatarY,
        'createdAt': (value.createdAt.toISOString()),
        'owner': value.owner,
    };
}