# webapp/src/shared/generated/models/UpdateChatDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * transcendence-app
 * The transcendence-app API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateChatDto
 */
export interface UpdateChatDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateChatDto
     */
    chatName?: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateChatDto
     */
    avatarX?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateChatDto
     */
    avatarY?: number;
}

/**
 * Check if a given object implements the UpdateChatDto interface.
 */
export function instanceOfUpdateChatDto(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function UpdateChatDtoFromJSON(json: any): UpdateChatDto {
    return UpdateChatDtoFromJSONTyped(json, false);
}

export function UpdateChatDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateChatDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'chatName': !exists(json, 'chatName') ? undefined : json['chatName'],
        'avatarX': !exists(json, 'avatarX') ? undefined : json['avatarX'],
        'avatarY': !exists(json, 'avatarY') ? undefined : json['avatarY'],
    };
}

export function UpdateChatDtoToJSON(value?: UpdateChatDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'chatName': value.chatName,
        'avatarX': value.avatarX,
        'avatarY': value.avatarY,
    };
}